Move test helpers to separate module.
